+++ packages/article-comments/__tests__/web/mocks.js
@@ -1,0 +1,4 @@
+import { mockUserState } from "@times-components/user-state";
+
+// eslint-disable-next-line import/prefer-default-export
+export const UserState = mockUserState();