+++ packages/ts-components/src/index.ts
@@ -66,5 +66,8 @@ export {
   AlgoliaSearchProvider,
   useAlgoliaSearch
 } from './helpers/algolia/AlgoliaSearchProvider';
+export { HiddenDiv } from './components/common-styles';
 
 export { InlineMessage } from './components/inline-message/InlineMessage';
+
+export { InlineDialog } from './components/inline-dialog/InlineDialog';