+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.7.58](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.7.57...@times-components/article-list@9.7.58) (2022-01-12)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.7.57](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.7.56...@times-components/article-list@9.7.57) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.7.56](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.7.55...@times-components/article-list@9.7.56) (2022-01-10)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.7.55](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.7.54...@times-components/article-list@9.7.55) (2022-01-07)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.7.54](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.7.53...@times-components/article-list@9.7.54) (2022-01-04)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.7.53](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.7.52...@times-components/article-list@9.7.53) (2022-01-04)
 
 **Note:** Version bump only for package @times-components/article-list